requirements_dev.txt
-r requirements.txt
bumpversion==0.5.3
wheel==0.26.0
watchdog==0.8.3
flake8==2.5.1
tox==2.3.1
coverage==4.0.3
Sphinx==1.3.3
mock==1.3.0
factory-boy==2.6.0

